refactor(host): extract WorkspaceTools readiness/support builders (workspacetools-god-class-decomposition)
Split WorkspaceTools.cs's readiness-report and support-bundle construction
logic into dedicated WorkspaceReadinessReportBuilder and
WorkspaceSupportBundleBuilder types (stateless sibling classes in the same
namespace), reducing the file from 1165 to 758 lines. Pure method relocation
plus Create/CreateWithoutTarget renames at the 6 in-file call sites; no public
API, MCP tool schema, or output-shape change.
